sqlalchemy_database/{{ cookiecutter.project_slug }}/requirements.txt
packaging
requests
PyMySQL
pandas
numpy
xlrd
openpyxl
ipykernel
notebook
SQLAlchemy
python-dotenv
setuptools
alembic
cookiecutter